feat(casper): gate rejected-deploy buffer population on finalization status
Catching-up validators replay historical blocks to get to the current
tip. For each block with non-empty body.rejected_deploys, the buffer-
population path extracts the rejected sigs' DeployData and adds them to
the local rejected-deploy buffer for re-proposal. Without a status
check, this admits sigs that have already been re-proposed and
finalized elsewhere in the chain, or sigs past their deployLifespan.

Two failure modes:

- Double-execution of already-finalized work. A rejected sig is added to
  the local buffer; on the validator's next proposal round, the buffer
  read includes the deploy; the new block contains the deploy; dedup
  picks the new proposal over the older finalized copy within merge
  scope; the merge produces a re-execution of canonical work against a
  different pre-state. Effects diverge. Consensus forks.

- Past-lifespan noise. The buffer read filter drops past-lifespan sigs
  at proposal time, but the entries still accumulate and churn through
  storage.

Fix: before admitting each sig to the buffer, run the deploy
finalization status resolver. Admit only if the state is Pending. Skip
Finalized / Failed / Expired — those sigs are terminally resolved in
the local canonical view and must not be re-proposed.

The gate is unconditional — not "catchup mode" flagged. A live merge
that re-emits a canonically-finalized sig would be equally unsafe; the
same gate defends against both.

Implementation:

- Extracted BlockAPI::deploy_finalization_status's algorithm into a
  pure function `deploy_finalization_status::resolve(dag, block_store,
  deploy_lifespan, sig)`. The async BlockAPI method now reduces to a
  thin wrapper that unwraps the engine cell and delegates. This makes
  the resolver callable from compute_parents_post_state without
  threading an EngineCell through the merge layer.

- Added should_admit_to_rejected_buffer helper in interpreter_util.rs
  that calls resolve and applies the admit rule. Conservative
  skip-on-error: transient storage failures skip the sig with a warn
  log; consistency errors skip with a warn log. Never admit on error —
  admit-on-error would reintroduce the double-execution bug under
  flaky storage.

- Wired the helper into compute_parents_post_state's buffer-populate
  block as a single predicate call, replacing the direct push.

Tests:
- Pure-resolver direct call: resolve_pure_function_returns_pending_for_unknown_sig
  verifies the extracted function is callable from a non-engine-cell
  context.

Deferred to later test work:
- Integration test exercising the gate-skips-finalized path (needs a
  fixture that produces merge rejection AND later finalization of the
  same sig — overlaps with equivocation + merge-rejection work).
- Full multi-node catchup simulation.

Consensus safety: the gate is a strict reduction of what enters the
buffer. Never adds sigs that weren't there; only drops sigs with a
terminal status in the current canonical view. Deterministic per
validator's DAG view.

Performance: O(deployLifespan) block reads per admit decision. For
typical rejection rates (0-3 per merge, lifespan ~50) this is sub-ms.
Full catchup of 1000 historical blocks with average 2 rejections each
adds ~100K block reads cumulatively — seconds of wall time.
